Created quickChatButtonsArray component + refactored
