+++ packages/app/src/game/module/board/overlays/quickChat/quickChatBox.tsx
@@ -1,4 +1,4 @@
-import { AppState } from "../../../../store";
+import { AppState } from "../../../../../store";
 import React from "react";
 import { useSelector } from "react-redux";
